+++ backend-spring/src/main/java/com/acmtrain/backend/dto/PointLogResponse.java
@@ -1,10 +1,12 @@
 package com.acmtrain.backend.dto;
 
+import java.math.BigDecimal;
+
 public record PointLogResponse(
         Long id,
         String sourceType,
         String reason,
-        Integer points,
+        BigDecimal points,
         String createdAt
 ) {
 }
