+++ backend-spring/src/main/java/com/acmtrain/backend/dto/RankingResponse.java
@@ -1,11 +1,13 @@
 package com.acmtrain.backend.dto;
 
+import java.math.BigDecimal;
+
 public record RankingResponse(
         Integer rankNo,
         String userName,
         Integer cfRating,
         Integer atcRating,
-        Integer totalPoints,
+        BigDecimal totalPoints,
         Integer solvedCount,
         Integer streakDays
 ) {
